+++ src/BDSExecOptions.cc
@@ -70,6 +70,7 @@ void BDSExecOptions::Parse(int argc, char **argv)
 {
   static struct option LongOptions[] = {{ "help" , 0, 0, 0 },
                                         { "version", 0, 0, 0 },
+                                        { "versionGit", 0, 0, 0 },
 					{ "verbose", 0, 0, 0 },
 					{ "verbose_step", 0, 0, 0 },         // old
 					{ "verboseSteppingBDSIM", 0, 0, 0 }, // new
@@ -158,6 +159,11 @@ void BDSExecOptions::Parse(int argc, char **argv)
 	    else if ( !strcmp(optionName, "version") )
 	      {
 		G4cout << "@BDSIM_VERSION@" << G4endl;
+		exit(0);
+	      }
+	    else if ( !strcmp(optionName, "versionGit") )
+	      {
+		G4cout << "@BDSIM_VERSION@ based on git commit SHA1: @VERSION_SHA1@" << G4endl;
 		exit(0);
 	      }
 	    else if ( !strcmp(optionName , "batch") )
@@ -429,6 +435,7 @@ void BDSExecOptions::Usage() const
 	<<"--verboseEventBDSIM          : display BDSIM information for every event "        << G4endl
 	<<"--verboseSteppingBDSIM       : display BDSIM tracking information after each step"<< G4endl
 	<<"--version                    : display version number"                            << G4endl
+	<<"--versionGit                 : display version number including git commit SHA1"  << G4endl
 	<<"--vis_debug                  : display all volumes in visualiser"                 << G4endl
 	<<"--vis_mac=<file>             : file with the visualisation macro script, default" << G4endl
 	<<"                               provided by BDSIM openGL (OGLSQt))"                << G4endl